requirements.txt
# NeuroScov Bot - Python Dependencies

# Core dependencies
python-dotenv==1.0.0          # Загрузка переменных окружения из .env

# Gemini AI
google-generativeai==0.8.3    # Google Gemini API

# Telegram
python-telegram-bot==21.7     # Telegram Bot API (async)
telethon==1.37.0              # Telegram User API для чтения каналов

# RSS Parsing
feedparser==6.0.11            # Парсинг RSS лент

# Database
# SQLite включен в стандартную библиотеку Python

# Scheduling
APScheduler==3.10.4           # Планировщик задач

# Image processing (опционально, для будущих версий)
Pillow==11.0.0                # Работа с изображениями

# Async support
aiohttp==3.11.7               # Асинхронные HTTP запросы
asyncio                       # Асинхронное программирование (встроено в Python 3.7+)

# Utilities
python-dateutil==2.9.0        # Работа с датами
pytz==2024.2                  # Временные зоны
